+++ rust/src/api/structs.rs
@@ -71,3 +71,6 @@ pub enum KitchenSink {
     Enums(Weekdays),
 }
 
+
+
+